requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg>=3.1.18

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
ldap3==2.9.1

# HTTP clients for integrations
httpx==0.25.1
aiohttp==3.9.0

# Task queue
celery==5.3.4
redis==5.0.1

# Utilities
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
